Moved tests related to ExternalContact and SelfContact to test_joint.py
